Refactor seeders and models namespaces in GestionRH directory
