--- app/app/Models/RH/Apprenant.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models\RH;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use App\Models\User;
-
-class Apprenant extends User
-{
-    use HasFactory;
-}
