--- app/app/Models/RH/Formateur.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Models\RH;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use App\Models\User;
-
-
-class Formateur extends User
-{
-    use HasFactory;
-}
