+++ app/database/seeders/GestionRHSeeder.php
@@ -4,15 +4,15 @@ namespace Database\Seeders;
 
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
-use Database\Seeders\RH\UserSeeder;
-class RhSeeder extends Seeder
+use Database\Seeders\GestionRH\UserSeeder;
+class GestionRHSeeder extends Seeder
 {
     /**
      * Run the database seeds.
      */
     public function run(): void
     {
-        $this->call(RhSeeder::Classes());
+        $this->call(GestionRHSeeder::Classes());
     }
